+++ NEWS.md
@@ -1,3 +1,4 @@
+- Add pad method for waves.
 - Improve sparse spectrum plotting.
 - Implement spectral filters.
 - Implement ADSR filter.